refactor: change Geometry conversions to extension functions
