+++ ground/src/main/java/com/google/android/ground/persistence/local/room/converter/LocationOfInterestMutationConverter.kt
@@ -35,7 +35,7 @@ object LocationOfInterestMutationConverter :
       surveyId = model.surveyId,
       jobId = model.jobId,
       type = MutationEntityType.fromMutationType(model.type),
-      newGeometry = model.geometry?.let { GeometryConverter.convertToDataStoreObject(it) },
+      newGeometry = model.geometry?.toLocalDataStoreObject(),
       userId = model.userId,
       locationOfInterestId = model.locationOfInterestId,
       syncStatus = MutationEntitySyncStatus.fromMutationSyncStatus(model.syncStatus),
@@ -52,7 +52,7 @@ object LocationOfInterestMutationConverter :
       surveyId = entity.surveyId,
       jobId = entity.jobId,
       type = entity.type.toMutationType(),
-      geometry = entity.newGeometry?.let { GeometryConverter.convertFromDataStoreObject(it) },
+      geometry = entity.newGeometry?.toModelObject(),
       userId = entity.userId,
       locationOfInterestId = entity.locationOfInterestId,
       syncStatus = entity.syncStatus.toMutationSyncStatus(),
